+++ src/main/java/com/gugucon/shopping/pay/dto/request/PayCreateRequest.java
@@ -1,11 +1,12 @@
-package com.gugucon.shopping.pay.dto;
+package com.gugucon.shopping.pay.dto.request;
 
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 public final class PayCreateRequest {
