+++ src/main/java/com/gugucon/shopping/pay/infrastructure/PayValidator.java
@@ -1,6 +1,6 @@
 package com.gugucon.shopping.pay.infrastructure;
 
-import com.gugucon.shopping.pay.dto.PayValidationRequest;
+import com.gugucon.shopping.pay.dto.request.PayValidationRequest;
 
 public interface PayValidator {
