+++ src/test/java/com/gugucon/shopping/integration/PayIntegrationTest.java
@@ -10,7 +10,11 @@ import com.gugucon.shopping.member.dto.request.LoginRequest;
 import com.gugucon.shopping.member.dto.request.SignupRequest;
 import com.gugucon.shopping.order.repository.OrderItemRepository;
 import com.gugucon.shopping.order.repository.OrderRepository;
-import com.gugucon.shopping.pay.dto.*;
+import com.gugucon.shopping.pay.dto.request.PayCreateRequest;
+import com.gugucon.shopping.pay.dto.request.PayValidationRequest;
+import com.gugucon.shopping.pay.dto.response.PayCreateResponse;
+import com.gugucon.shopping.pay.dto.response.PayInfoResponse;
+import com.gugucon.shopping.pay.dto.response.PayValidationResponse;
 import com.gugucon.shopping.pay.repository.PayRepository;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
